requirements.txt
# Core dependencies for Agentic Graph RAG
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
networkx>=3.1
neo4j>=5.14.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
